WIP: chore/tests: Share get_bin fo benching
