+++ resources/test-scripts/install-vault-and-authd
@@ -1,0 +1,93 @@
+#!/bin/bash
+
+# Run a baby-fleming network
+set -e -x
+
+[[ -z $SAFE_AUTHD_PATH ]] && export SAFE_AUTHD_PATH=~/safe-authd
+[[ -z $SAFE_VAULT_PATH ]] && export SAFE_VAULT_PATH=~/safe-vault
+
+safe_vault_version=$1
+uname_output=$(uname -a)
+authd_bin_name="safe-authd"
+case $uname_output in
+    MSYS_NT*)   # Windows in Jenkins
+        authd_bin_name="safe-authd.exe"
+        ;;
+    MINGW*)     # Windows in GH Actions
+        authd_bin_name="safe-authd.exe"
+        ;;
+esac
+
+function download_safe_vault() {
+    rm -rf $SAFE_VAULT_PATH
+    mkdir $SAFE_VAULT_PATH
+    cwd=$(pwd)
+    cd $SAFE_VAULT_PATH
+
+    # TODO: If no version specified just use that installed by "safe vault install"
+
+    vault_bin_name="safe_vault"
+    case $uname_output in
+        Linux*)
+            curl -L -O "https://github.com/maidsafe/safe_vault/releases/download/$vault_version/safe_vault-$vault_version-x86_64-unknown-linux-musl.tar.gz"
+            tar xvf "safe_vault-$vault_version-x86_64-unknown-linux-musl.tar.gz"
+            rm "safe_vault-$vault_version-x86_64-unknown-linux-musl.tar.gz"
+            ;;
+        Darwin*)
+            curl -L -O "https://github.com/maidsafe/safe_vault/releases/download/$vault_version/safe_vault-$vault_version-x86_64-apple-darwin.tar.gz"
+            tar xvf "safe_vault-$vault_version-x86_64-apple-darwin.tar.gz"
+            rm "safe_vault-$vault_version-x86_64-apple-darwin.tar.gz"
+            ;;
+        MSYS_NT* | MINGW*)  # Windows in Jenkins and in GH Actions respectively
+            curl -L -O "https://github.com/maidsafe/safe_vault/releases/download/$vault_version/safe_vault-$vault_version-x86_64-pc-windows-gnu.tar.gz"
+            tar xvf "safe_vault-$vault_version-x86_64-pc-windows-gnu.tar.gz"
+            rm "safe_vault-$vault_version-x86_64-pc-windows-gnu.tar.gz"
+            vault_bin_name="safe_vault.exe"
+            ;;
+        *)
+            echo "Platform not supported. Please extend to support this platform."
+            exit 1
+    esac
+
+    cd $cwd
+}
+
+
+function run_safe_vault() {
+    cwd=$(pwd)
+    cd $SAFE_VAULT_PATH
+
+    vault_bin_name="safe_vault"
+
+    echo "Launching local safe_vault..."
+    ./$vault_bin_name &
+
+    cd $cwd
+}
+
+
+function build_safe_authd() {
+    cd safe-authd
+    cargo build --release
+
+    rm -rf $SAFE_AUTHD_PATH
+    mkdir $SAFE_AUTHD_PATH
+    mv ../target/release/$authd_bin_name $SAFE_AUTHD_PATH
+    pwd
+    cd -
+}
+
+function run_safe_authd() {
+    cd $SAFE_AUTHD_PATH
+
+    # this won't have any effect on Linux/Mac, but it's required for Windows
+    ./$authd_bin_name install || true
+
+    echo "Stopping safe-authd daemon.."
+    ./$authd_bin_name stop || true
+
+    echo "Launching safe-authd daemon..."
+    ./$authd_bin_name start
+    sleep 15
+    cd -
+}
