+++ safe-cli/tests/common/mod.rs
@@ -15,25 +15,13 @@ use std::collections::BTreeMap;
 use std::{env, str::FromStr};
 use unwrap::unwrap;
 
+pub use safe_utilities::get_bin_location;
+
 #[allow(dead_code)]
 pub const CLI: &str = "safe";
 #[allow(dead_code)]
 pub const SAFE_PROTOCOL: &str = "safe://";
 const TEST_FOLDER: &str = "../testdata/";
-
-#[allow(dead_code)]
-pub fn get_bin_location() -> String {
-    let target_dir = match env::var("CARGO_TARGET_DIR") {
-        Ok(target_dir) => target_dir,
-        Err(_) => "../target".to_string(),
-    };
-
-    if cfg!(debug_assertions) {
-        format!("{}{}", target_dir, "/debug/safe")
-    } else {
-        format!("{}{}", target_dir, "/release/safe")
-    }
-}
 
 #[allow(dead_code)]
 pub fn create_preload_and_get_keys(preload: &str) -> (String, String) {
