+++ utilities/Cargo.toml
@@ -1,0 +1,52 @@
+[package]
+name = "safe-utilities"
+version = "0.1.0"
+description = "SAFE CLI Test and benchmark utilities"
+authors = ["bochaco <gabrielviganotti@gmail.com>", "Josh Wilson <joshuef@gmail.com>"]
+publish = false
+license = "MIT OR BSD-3-Clause"
+repository = "https://github.com/maidsafe/safe-api"
+edition = "2018"
+
+
+[dependencies]
+# async-std = "1.4.0"
+# crossbeam = "0.7.3"
+# directories = "~2.0.1"
+env_logger = "0.6.0"
+envy = "0.4.0"
+human-panic = "1.0.1"
+log = "0.4.6"
+# prettytable-rs = "^0.8"
+rand = "0.6.5"
+relative-path = "0.4.0"
+# reqwest = "0.9.22"
+# rpassword = "~4.0.5"
+# safe-nlt = { git = "https://github.com/maidsafe/safe-network-launch-tool.git" }
+# serde = "1.0.91"
+# serde_json = "1.0.39"
+# serde_yaml = "0.8.11"
+# shrust = "0.0.7"
+# structopt = "~0.3.5"
+# tokio = { version = "^0.2.13", features = ["macros"] }
+# pretty-hex = "0.1.1"
+# ansi_term = "0.9.0"
+# isatty = "0.1"
+# safe-nd = "^0.8.0"
+
+[features]
+mock-network = ["safe-api/mock-network"]
+fake-auth = ["safe-api/fake-auth"]
+scl-mock = ["safe-api/scl-mock"]
+
+[dependencies.safe-api]
+path = "../safe-api"
+default-features = false
+features = ["app", "authd_client"]
+
+[dev-dependencies]
+# assert_cmd = "0.11.1"
+# duct = "0.12.0"
+# predicates = "1.0.0"
+# pretty_assertions = "0.6.1"
+# unwrap = "~1.2.1"
